# contract-test/openapi.yaml
openapi: "3.0.3"
info:
  title: VSP Security Platform API
  version: "2.0.0"
  description: |
    VSP (Vulnerability Scanning Platform) — Enterprise Security API.

    Authentication: Bearer JWT token từ POST /api/v1/auth/login.

    ## Versioning
    - **2.0.0** — Synced with frontend reality (Phase T2). Adds P4 compliance,
      Zero Trust, RMF/ATO, SOAR, threat intel, UEBA, governance, log pipeline,
      remediation, OSCAL exports, scheduler, assets, AI analyst, vuln management.
    - **1.0.0** — Initial spec. Core auth + scan + admin only.

    ## Contract Testing
    This spec is enforced by `vsp_contract_test.js`. Any drift between
    frontend calls and this document fails CI.
  contact:
    name: VSP Security Team
    email: soc@agency.gov
  license:
    name: Proprietary
servers:
  - url: http://localhost:8921
    description: Development
  - url: http://127.0.0.1:8921
    description: Local loopback
  - url: https://vsp.agency.gov
    description: Production
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
...
paths:
...
  /api/p4/rmf/task:
    post:
      summary: Update RMF task status
      tags: [RMF]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [task_id, status]
              properties:
                task_id: { type: string }
                status:  { type: string, enum: [pending, in_progress, complete] }
      responses:
        "200": { description: Updated }
...